FileFilter

+++ FileFiltrerCSharp/FileFiltrerCSharp/FileMover.cs
@@ -24,7 +24,10 @@ namespace FileFiltrerCSharp
                     string destFile = Path.Combine(targetPath, fileName);
                     Console.WriteLine("Déplacement de '" + fileName + "' dans '" + targetPath );
 
-                    File.Move(sourcefile, destFile);
+                    if (!Directory.Exists(targetPath))
+                    {
+                        File.Move(sourcefile, destFile);
+                    }
                 }
             }
         }
